functions/requirements.txt
# Core
firebase-functions
firebase-admin
google-cloud-firestore
Flask
Werkzeug
flask-cors
# Optional (commented out if not used directly)
pandas
openpyxl
